scanline adjustment and interpret last command line argument as int instead of float.

+++ MRIUI.h
@@ -73,7 +73,7 @@ class MRIUI : public QMainWindow {
     IMouse* mouse;
 
     bool useCPU;
-    float phantomSize;
+    unsigned int  phantomSize;
 
     MRISim* sim;
     IMRIKernel* kern;
@@ -108,7 +108,7 @@ public slots:
     void SetSimView(bool toggle);
     void SetRFView(bool toggle);
 public:
-    MRIUI(OpenEngine::Display::QtEnvironment *env, bool, float);    
+    MRIUI(OpenEngine::Display::QtEnvironment *env, bool, unsigned int);    
 
 };
